pyproject.toml
[tool.poetry]
name = "mysite"
version = "0.1.0"
description = ""
authors = ["Nikolai <nicolas.gordienko@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "~4.1.0"
django-bootstrap4 = "^23.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
dj-database-url = "^2.0.0"
whitenoise = "^6.5.0"
weasyprint = "^59.0"
requests = "^2.31.0"
django-ckeditor = "^6.7.0"
pdfkit = "^1.0.0"
drf-yasg = "^1.21.7"
djangorestframework = "^3.14.0"
readability-lxml = "^0.8.1"
django-debug-toolbar = "^4.3.0"
django-cors-headers = "^4.3.1"
langdetect = "^1.0.9"
newspaper3k = "^0.2.8"
openai = "^1.12.0"
django-modeltranslation = "~0.18.9"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
django-test = "^0.4030"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
